# Fix length of allocated slices in SFXTagsToJaegerTags