There are only 3 possible tags in processTags and N+4 in total. Allocate an N+4 slice and fill spanTags initially then processTags.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkSFXTagsToJaegerTags-16     1178          753           -36.08%

benchmark                           old allocs     new allocs     delta
BenchmarkSFXTagsToJaegerTags-16     2              1              -50.00%

benchmark                           old bytes     new bytes     delta
BenchmarkSFXTagsToJaegerTags-16     3840          2304          -40.00%
```