@bogdandrutu 2020-06-16
Travis build is caused by some coverage of a function that I did not touch.